requirements.txt
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
tensorflow==2.15.0
Pillow==10.0.1
numpy==1.26.2
python-dotenv==1.0.0
WTForms==3.1.1
